# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinic/condition-prices:
    get:
      operationId: getConditionPrices
      tags: [dental]
      summary: Get clinic condition prices and ICD-10 codes
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Condition prices map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionPricesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateConditionPrices
      tags: [dental]
      summary: Update clinic condition prices (owner/admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConditionPricesRequest"
      responses:
        "200":
          description: Condition prices updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionPricesResponse"
        "403":
          description: Forbidden - requires owner or admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{id}/teeth:
    get:
      operationId: listTeeth
      tags: [dental]
      summary: Get all tooth records for a patient
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tooth records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeethResponse"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /patients/{id}/treatment-plans:
    get:
      operationId: listTreatmentPlans
      tags: [treatment-plans]
      summary: List all treatment plans for a patient
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of treatment plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlansResponse"
  /patients/{id}/treatment-plan:
    get:
      operationId: getActiveTreatmentPlan
      tags: [treatment-plans]
      summary: Get active (non-completed) treatment plan for a patient
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active treatment plan or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanResponse"
    post:
      operationId: createTreatmentPlan
      tags: [treatment-plans]
      summary: Create a treatment plan (auto-generates items from tooth records if items not provided)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatmentPlanRequest"
      responses:
        "201":
          description: Created treatment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanResponse"
  /patients/{id}/treatment-plan/{planId}:
    patch:
      operationId: updateTreatmentPlan
      tags: [treatment-plans]
      summary: Update treatment plan notes or status
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreatmentPlanRequest"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanBaseResponse"
  /patients/{id}/treatment-plan/{planId}/approve:
    post:
      operationId: approveTreatmentPlan
      tags: [treatment-plans]
      summary: Approve treatment plan (sets status to approved)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: planId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanBaseResponse"
  /patients/{id}/treatment-plan/{planId}/items:
    post:
      operationId: addTreatmentPlanItem
      tags: [treatment-plans]
      summary: Add an item to a treatment plan
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTreatmentPlanItemRequest"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanItemResponse"
  /patients/{id}/treatment-plan/{planId}/items/{itemId}:
    patch:
      operationId: updateTreatmentPlanItem
      tags: [treatment-plans]
      summary: Update a treatment plan item (title, price, status)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: planId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreatmentPlanItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentPlanItemResponse"
  /patients/{id}/treatment-plan/{planId}/items/{itemId}/complete:
    post:
      operationId: completeTreatmentPlanItem
      tags: [treatment-plans]
      summary: Mark item as completed — creates a billing procedure
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: planId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed item and procedure id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteTreatmentPlanItemResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
      required:
        - success
        - error
...
    ToothCondition:
      type: string
      enum:
        - healthy
        - cavity
        - treated
        - crown
        - root_canal
        - implant
        - missing
        - extraction_needed
    ToothRecord:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        toothFdi:
          type: integer
        condition:
          $ref: "#/components/schemas/ToothCondition"
        notes:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          nullable: true
      required:
        - id
        - clinicId
        - patientId
        - toothFdi
        - condition
        - updatedAt
...
    TeethResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            teeth:
              type: array
              items:
                $ref: "#/components/schemas/ToothRecord"
          required:
            - teeth
      required:
        - success
        - data
    ConditionPriceEntry:
      type: object
      properties:
        price:
          type: number
        mkb10:
          type: string
      required:
        - price
        - mkb10
    ConditionPricesMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/ConditionPriceEntry"
      description: Map of condition name to price and ICD-10 code
    ConditionPricesResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            prices:
              $ref: "#/components/schemas/ConditionPricesMap"
          required:
            - prices
      required:
        - success
        - data
    UpdateConditionPriceItem:
      type: object
      properties:
        price:
          type: number
          minimum: 0
        mkb10Code:
          type: string
      required:
        - price
    UpdateConditionPricesRequest:
      type: object
      properties:
        prices:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/UpdateConditionPriceItem"
          description: Map of condition name to price and optional МКБ-10 code
      required:
        - prices
...
    TreatmentPlanStatus:
      type: string
      enum: [draft, approved, in_progress, completed, cancelled]
    TreatmentPlanItemStatus:
      type: string
      enum: [pending, completed, cancelled]
    TreatmentPlanItem:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        toothFdi:
          type: integer
          nullable: true
        condition:
          type: string
          nullable: true
        mkb10Code:
          type: string
          nullable: true
        title:
          type: string
        price:
          type: number
        status:
          $ref: "#/components/schemas/TreatmentPlanItemStatus"
        sortOrder:
          type: integer
        procedureId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - planId
        - clinicId
        - patientId
        - title
        - price
        - status
        - sortOrder
        - createdAt
    TreatmentPlan:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
          nullable: true
        planNumber:
          type: integer
        status:
          $ref: "#/components/schemas/TreatmentPlanStatus"
        notes:
          type: string
          nullable: true
        totalCost:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/TreatmentPlanItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - patientId
        - planNumber
        - status
        - totalCost
        - items
        - createdAt
        - updatedAt
    TreatmentPlanResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            plan:
              oneOf:
                - $ref: "#/components/schemas/TreatmentPlan"
                - type: "null"
          required:
            - plan
      required:
        - success
        - data
    TreatmentPlanBaseResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            plan:
              $ref: "#/components/schemas/TreatmentPlan"
          required:
            - plan
      required:
        - success
        - data
    TreatmentPlansResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            plans:
              type: array
              items:
                $ref: "#/components/schemas/TreatmentPlan"
          required:
            - plans
      required:
        - success
        - data
    TreatmentPlanItemResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            item:
              $ref: "#/components/schemas/TreatmentPlanItem"
          required:
            - item
      required:
        - success
        - data
    CompleteTreatmentPlanItemResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            item:
              $ref: "#/components/schemas/TreatmentPlanItem"
            procedureId:
              type: string
          required:
            - item
            - procedureId
      required:
        - success
        - data
    CreateTreatmentPlanRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              toothFdi:
                type: integer
              condition:
                type: string
              mkb10Code:
                type: string
              title:
                type: string
              price:
                type: number
            required:
              - title
              - price
    UpdateTreatmentPlanRequest:
      type: object
      properties:
        notes:
          type: string
          nullable: true
    AddTreatmentPlanItemRequest:
      type: object
      properties:
        toothFdi:
          type: integer
        condition:
          type: string
        mkb10Code:
          type: string
        title:
          type: string
        price:
          type: number
      required:
        - title
        - price
    UpdateTreatmentPlanItemRequest:
      type: object
      properties:
        title:
          type: string
        price:
          type: number
        sortOrder:
          type: integer
          minimum: 0
        status:
          type: string
          enum:
            - cancelled